+++ src/components/addProduct/helpers/index.js
@@ -1,2 +1,3 @@
 export {FileInput} from './fileInput'
 export {ImageDisplay} from './imageDisplay'
+export {ProductInfo} from './productInfo'
